### 2018-12-13 #1126 @whedon
[PRE REVIEW]: arcesetc: ARC Echelle Spectroscopic Exposure Time Calculator
**Submitting author:** @bmorris3 (<a href="http://orcid.org/0000-0003-2528-3409">Brett Morris</a>)
**Repository:** <a href="https://github.com/bmorris3/arcesetc" target ="_blank">https://github.com/bmorris3/arcesetc</a>
**Version:** 0.1
