+++ notebooks/target_paths.txt
@@ -1,3 +1,5 @@
+/Users/bmmorris/data/arcesetc/allison/HD289002_ARCES_120s.0094.wfrmcpc.fits
+/Users/bmmorris/data/arcesetc/allison/Sirius_ARCES_halfsec.0091.wfrmcpc.fits
 /Users/bmmorris/data/arcesetc/yuta/HR3454.0025.wfrmcpc.fits
 /Users/bmmorris/data/arcesetc/yuta/HR4963.0030.wfrmcpc.fits
 /Users/bmmorris/data/arcesetc/yuta/HR5191.0002.wfrmcpc.fits
