### 2018-12-14 #1126 @arfon
> May I recommend @nespinoza or @wtgee as reviewers? Thanks to each for considering.

Great suggestions. @nespinoza or @wtgee - would you be interested in reviewing this submission for JOSS?

Also, @bmorris3 - can you updated your citation to the published version of the Astropy 2.0 paper? http://adsabs.harvard.edu/cgi-bin/nph-bib_query?bibcode=2018arXiv180102634T&data_type=BIBTEX&db_key=PRE&nocookieset=1
