### 2018-12-14 b9adffb
Updating astropy reference
