+++ docs/paper.bib
@@ -1,59 +1,56 @@
 @ARTICLE{astropy:2018,
-       author = {{The Astropy Collaboration} and {Price-Whelan}, A.~M. and
-        {Sip{\'{o}}cz}, B.~M. and {G{\"u}nther}, H.~M. and {Lim}, P.~L.
-        and {Crawford}, S.~M. and {Conseil}, S. and {Shupe}, D.~L. and
-        {Craig}, M.~W. and {Dencheva}, N. and {Ginsburg}, A. and
-        {VanderPlas}, J.~T. and {Bradley}, L.~D. and {P{\'e}rez-
-        Su{\'a}rez}, D. and {de Val-Borro}, M. and {Aldcroft}, T.~L. and
-        {Cruz}, K.~L. and {Robitaille}, T.~P. and {Tollerud}, E.~J. and
-        {Ardelean}, C. and {Babej}, T. and {Bachetti}, M. and {Bakanov},
-        A.~V. and {Bamford}, S.~P. and {Barentsen}, G. and {Barmby}, P.
-        and {Baumbach}, A. and {Berry}, K.~L. and {Biscani}, F. and
-        {Boquien}, M. and {Bostroem}, K.~A. and {Bouma}, L.~G. and
-        {Brammer}, G.~B. and {Bray}, E.~M. and {Breytenbach}, H. and
-        {Buddelmeijer}, H. and {Burke}, D.~J. and {Calderone}, G. and
-        {Cano Rodr{\'\i}guez}, J.~L. and {Cara}, M. and {Cardoso},
-        J.~V.~M. and {Cheedella}, S. and {Copin}, Y. and {Crichton}, D.
-        and {D{\'A}vella}, D. and {Deil}, C. and {Depagne}, {\'E}. and
-        {Dietrich}, J.~P. and {Donath}, A. and {Droettboom}, M. and
-        {Earl}, N. and {Erben}, T. and {Fabbro}, S. and {Ferreira},
-        L.~A. and {Finethy}, T. and {Fox}, R.~T. and {Garrison}, L.~H.
-        and {Gibbons}, S.~L.~J. and {Goldstein}, D.~A. and {Gommers}, R.
-        and {Greco}, J.~P. and {Greenfield}, P. and {Groener}, A.~M. and
-        {Grollier}, F. and {Hagen}, A. and {Hirst}, P. and {Homeier}, D.
-        and {Horton}, A.~J. and {Hosseinzadeh}, G. and {Hu}, L. and
-        {Hunkeler}, J.~S. and {Ivezi{\'c}}, {\v{Z}}. and {Jain}, A. and
-        {Jenness}, T. and {Kanarek}, G. and {Kendrew}, S. and {Kern},
-        N.~S. and {Kerzendorf}, W.~E. and {Khvalko}, A. and {King}, J.
-        and {Kirkby}, D. and {Kulkarni}, A.~M. and {Kumar}, A. and
-        {Lee}, A. and {Lenz}, D. and {Littlefair}, S.~P. and {Ma}, Z.
-        and {Macleod}, D.~M. and {Mastropietro}, M. and {McCully}, C.
-        and {Montagnac}, S. and {Morris}, B.~M. and {Mueller}, M. and
-        {Mumford}, S.~J. and {Muna}, D. and {Murphy}, N.~A. and
-        {Nelson}, S. and {Nguyen}, G.~H. and {Ninan}, J.~P. and
-        {N{\"o}the}, M. and {Ogaz}, S. and {Oh}, S. and {Parejko}, J.~K.
-        and {Parley}, N. and {Pascual}, S. and {Patil}, R. and {Patil},
[…]
