### 2018-12-14 #1126 @bmorris3
@wtgee That's perfect thank you! Also, we should totally meet up while you're in Seattle, I'd love to meet you in person 😄
