### 2018-12-14 #1130 @arfon
Hi @wtgee, thanks again for agreeing to review this submission for JOSS. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Any questions/concerns please let me know. Also, I realize the holidays and AAS are coming fast so could you aim to have your review complete by the 21st January?
