### 2018-12-14 #1126 @nespinoza
> @wtgee That's perfect thank you! Also, we should totally meet up while you're in Seattle, I'd love to meet you in person 😄

Got late to the party but glad to see @wtgee will review this! Good luck with the submission @bmorris3!
