### 2018-12-19 #1130 @bmorris3
@wtgee – Looking over the review criteria, I realize that the "statement of need" might not be obvious in my write-up without some context. It should be noted that ARCES does not currently have any exposure time calculator, so a tool like `arcestec` is very much needed.
