### 2019-01-31 c4d0c1b
added trailing periods to param doc sentences. Removed trailing 'If'
