### 2019-02-01 75212a1
Update contributing.rst
