### 2019-02-01 d18fa37
Fixing issue #4
