+++ setup.cfg
@@ -38,7 +38,7 @@ url = https://github.com/bmorris3/arcesetc
 edit_on_github = False
 github_project = bmorris3/arcesetc
 # install_requires should be formatted as a comma-separated list, e.g.:
-install_requires = astropy, numpy, matplotlib, h5py
+install_requires = astropy, numpy, matplotlib, h5py, pytest-astropy
 # version should be PEP440 compatible (https://www.python.org/dev/peps/pep-0440/)
 version = 0.2.dev
 # Note: you will also need to change this in your package's __init__.py
