### 2019-02-01 5ac33ed
Clarifications for #10
