### 2019-02-01 754d0ed
Update for Issue #6
