### 2019-02-01 #1130 @bmorris3
Thanks @wtgee for the issues! We've addressed each one.
